+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Log the endpoint when a quic connection can't be established.
+
 ## [0.3.2] - 2024-01-31
 
 ### Added